Update Aircraft_Noise_Emissions.py
Updated lines 410 to 480 to only sum results when they cross the penalty time. Otherwise it multiplies the timestep by the number of flights in that timeframe (ex. multiplies the timesteps for daylight hours by the number of operations conducted in the daylight timeframe). This does not include the update needed for variable timesteps!